# src/docs/swagger.yaml
definitions:
  api_handlers.personData:
    properties:
      first_name:
        maxLength: 10
        minLength: 4
        type: string
      last_name:
        maxLength: 20
        minLength: 6
        type: string
      mobile_number:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - first_name
    - last_name
    - mobile_number
    type: object
...
  constants.NotificationMessageType:
    enum:
    - info
    - warn
    - error
    - success
    type: string
    x-enum-varnames:
    - INFO
    - WARN
    - ERROR
    - SUCCESS
...
  github_com_jaganathanb_dapps-api_api_dto.NotificationsPayload:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      deletedAt:
        type: string
      deletedBy:
        type: integer
      id:
        type: integer
      isRead:
        type: boolean
      message:
        type: string
      messageType:
        $ref: '#/definitions/constants.NotificationMessageType'
      modifiedAt:
        type: string
      modifiedBy:
        type: integer
      title:
        type: string
      userId:
        type: integer
    type: object
...
  github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_validations.ValidationError'
        type: array
    type: object
  github_com_jaganathanb_dapps-api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  github_com_jaganathanb_dapps-api_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v{version}/notifications:
    delete:
      consumes:
      - application/json
      description: Delete notifications from GST Web
      parameters:
      - description: UserId
        in: header
        name: dapps-user-id
        required: true
        type: integer
      - default: 1
        description: Version
        enum:
        - 1
        - 2
        in: path
        name: version
        required: true
        type: integer
      - description: NotificationsPayload
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_dto.NotificationsPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete notifications
      tags:
      - Notifications
    get:
      consumes:
      - application/json
      description: Notifications for GST Web
      parameters:
      - description: UserId
        in: header
        name: dapps-user-id
        required: true
        type: integer
      - default: 1
        description: Version
        enum:
        - 1
        - 2
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get notifications
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      description: Add notifications for GST Web
      parameters:
      - description: UserId
        in: header
        name: dapps-user-id
        required: true
        type: integer
      - default: 1
        description: Version
        enum:
        - 1
        - 2
        in: path
        name: version
        required: true
        type: integer
      - description: NotificationsPayload
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_dto.NotificationsPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Add notifications
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: Update notifications for GST Web
      parameters:
      - description: UserId
        in: header
        name: dapps-user-id
        required: true
        type: integer
      - default: 1
        description: Version
        enum:
        - 1
        - 2
        in: path
        name: version
        required: true
        type: integer
      - description: NotificationsPayload
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_dto.NotificationsPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update notifications
      tags:
      - Notifications
...
  /v1/test:
    get:
      consumes:
      - application/json
      description: Test
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      summary: Test
      tags:
      - Test
  /v1/test/binder/body:
    post:
      consumes:
      - application/json
      description: BodyBinder
      parameters:
      - description: person data
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/api_handlers.personData'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
            - properties:
                validationErrors:
                  type: object
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: BodyBinder
      tags:
      - Test
  /v1/test/binder/uri/{id}/{name}:
    post:
      consumes:
      - application/json
      description: BodyBinder
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: user name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
            - properties:
                validationErrors:
                  type: object
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: BodyBinder
      tags:
      - Test
  /v1/test/user/{id}:
    get:
      consumes:
      - application/json
      description: UserById
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_jaganathanb_dapps-api_api_helper.BaseHttpResponse'
      summary: UserById
      tags:
      - Test
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"